% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtraking.R
\name{backtraking}
\alias{backtraking}
\title{matrice de sudoku incomplète, nombre de tour maximal}
\usage{
backtraking(A, x)
}
\value{
solution sudoku et nombre de tours dans l'algorithme
}
\description{
Algorithme de Backtracking
}
