% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emplacementdisponible.R
\name{emplacementdisponible}
\alias{emplacementdisponible}
\title{Vérifie si k peut être inséré à l'emplacement (i,j).}
\usage{
emplacementdisponible(i, j, k, A)
}
\value{
un booléen
}
\description{
Vérifie si k peut être inséré à l'emplacement (i,j).
}
