% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existesurbloc.R
\name{existesurbloc}
\alias{existesurbloc}
\title{Vérifie si k peut être insérer dans le bloc correspondant à la case (i,j)}
\usage{
existesurbloc(i, j, k, A)
}
\value{
un booléen
}
\description{
Vérifie si k peut être insérer dans le bloc correspondant à la case (i,j)
}
