% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existesurcolonne.R
\name{existesurcolonne}
\alias{existesurcolonne}
\title{Vérifie si k se trouve sur la colonne j.}
\usage{
existesurcolonne(j, k, A)
}
\value{
un booléen
}
\description{
Vérifie si k se trouve sur la colonne j.
}
