% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grilleincomplete.R
\name{grilleincomplete}
\alias{grilleincomplete}
\title{Génère le sudoku incomplet à remplir.}
\usage{
grilleincomplete(m)
}
\value{
une liste contenant le sudoku à remplir, la solution, et le nombre de tours dans l'algorithme de backtracking.
}
\description{
Génère le sudoku incomplet à remplir.
}
