% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rang.R
\name{rang}
\alias{rang}
\title{Détermine la position des cases incomplètes dans l'algorithme de Backtracking.}
\usage{
rang(A)
}
\value{
Une liste de position.
}
\description{
Détermine la position des cases incomplètes dans l'algorithme de Backtracking.
}
