% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressioncase.R
\name{suppressioncase}
\alias{suppressioncase}
\title{Supprime une case aléatoire du sudoku A.}
\usage{
suppressioncase(A)
}
\value{
Le sudoku A avec une case en moins.
}
\description{
Supprime une case aléatoire du sudoku A.
}
